[MIG] base_export_manager: Upgrade to v9 * Upgrade base_export_manager to v9 * Rename JS to base_export_manager
