+++ base_export_manager/models/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# © 2015 Antiun Ingeniería S.L. - Antonio Espinosa
+# Copyright 2015 Antiun Ingeniería S.L. - Antonio Espinosa
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 from . import ir_exports, ir_exports_line
