+++ base_export_manager/tests/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
-# © 2015 Antiun Ingeniería S.L. - Javier Iniesta
-# © 2015 Antiun Ingeniería S.L. - Jairo Llopis
+# Copyright 2015 Antiun Ingeniería S.L. - Javier Iniesta
+# Copyright 2015 Antiun Ingeniería S.L. - Jairo Llopis
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 from . import test_ir_exports, test_ir_exports_line
