+++ base_export_manager/tests/test_ir_exports_line.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# © 2015 Antiun Ingenieria S.L. - Javier Iniesta
+# Copyright 2015 Antiun Ingenieria S.L. - Javier Iniesta
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 from openerp.tests.common import TransactionCase
