+++ base_export_manager/views/ir_exports_view.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <!-- Copyright 2016 Jairo Llopis <jairo.llopis@tecnativa.com>
      License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
-<openerp>
-<data>
+
+<odoo>
 
 <record id="ir_exports_action" model="ir.actions.act_window">
     <field name="name">Export Profiles</field>
@@ -81,5 +81,4 @@
     </field>
 </record>
 
-</data>
-</openerp>
+</odoo>
